Moved start screen to its own file, removed return null after welcome is false to start game scene
