+++ Images.js
@@ -14,9 +14,9 @@ const Images = {
   rock1: require("./assets/background/rock1.png"),
   rock2: require("./assets/background/rock2.png"),
   rock3: require("./assets/background/rock3.png"),
-  smallRock: require("./assets/background/smallRock.png"),
-  bigRock: require("./assets/background/bigRock.png"),
-  tallRock: require("./assets/background/tallRock.png"),
+  smallRock: require("./assets/background/smallrock.png"),
+  bigRock: require("./assets/background/bigrock.png"),
+  tallRock: require("./assets/background/tallrock.png"),
   pause: require("./assets/pause.png"),
   jumpPaw: require("./assets/jumpPaw.png"),
   restart: require("./assets/restart.png"),
